refactor: migrate pr_loop, serve to drift-mcp (ADR-102)
(cherry picked from commit f497f3e265ddb71291e9af6d5bef04d9881145b9)
