+++ packages/drift-mcp/src/drift_mcp/pr_loop/__init__.py
@@ -1,0 +1,7 @@
+﻿# PR Review Loop — public API
+from drift_config._loader import PrLoopConfig  # noqa: F401
+
+from drift_mcp.pr_loop._engine import loop_until_approved  # noqa: F401
+from drift_mcp.pr_loop._models import LoopExitStatus, LoopState, ReviewState  # noqa: F401
+
+__all__ = ["loop_until_approved", "LoopState", "PrLoopConfig"]
