+++ packages/drift-mcp/src/drift_mcp/serve/__init__.py
@@ -1,0 +1,6 @@
+"""drift.serve — A2A-compatible HTTP server for drift.
+
+Optional module.  Requires ``pip install drift-analyzer[serve]``.
+"""
+
+from __future__ import annotations
